Updates for ShinobiCharts 2.7 etc to solve #1
* Updated to modern objective-C syntax
* Fixed libraries to use installed version of ShinobiCharts
* Updated to use latest version of ShinobiCharts

+++ PieChart/Datasource.h
@@ -32,6 +32,6 @@
 
 @property (nonatomic) NSString *selectedOS;
 
-- (int)selectedSliceIndex;
+- (NSInteger)selectedSliceIndex;
 
 @end
